refactor : OrderProcessingController의 명확성과 모듈성 향상을 위한 리팩토링
- OrderProcessingController 내의 orderProcess 메소드를 간소화하기 위해 요청 파라미터 처리 로직을 OrderProcessingService로 추출함.
- 주문 양식 데이터를 캡슐화하는 OrderFormDataDTO를 도입하여 데이터 처리와 가독성을 개선함.
- 주문 상세 정보 및 결제 처리 로직을 OrderProcessingService 내의 별도 메소드로 분리하여 관심사의 분리를 개선함.
- 코드 정리 및 유지 관리 용이성을 위한 메소드 추출을 활용함.
- 컨트롤러와 서비스 계층 간의 데이터 전송을 위해 일관된 명명 규칙을 사용하고 DTO를 활용함.
